+++ defines/docker.go
@@ -23,6 +23,7 @@ type DockerWrapper struct {
 	CreateExec       func(docker.CreateExecOptions) (*docker.Exec, error)
 	StartExec        func(string, docker.StartExecOptions) error
 	Ping             func() error
+	Stats            func(docker.StatsOptions) error
 }
 
 func NewDocker(endpoint, cert, key, ca string) *DockerWrapper {
